build: apply C23 standard with strict static analysis
- Enable C23 standard (-std=c23) with comprehensive warning flags
- Add strict type checking: -Wconversion, -Wsign-conversion
- Enable prototype checking: -Wmissing-prototypes, -Wstrict-prototypes
- Add security flags: -fstack-protector-strong
- Fix sign conversion warnings in common.h, input.c, server.c, client.c
- Fix jsmn.h parser to be C23 compliant with explicit casts
- Remove unused tokenize_with_quotes function
- Update Makefile to use clang with C23 flags by default

All code now compiles with zero warnings under strict C23 mode.

+++ src/common/common.h
@@ -24,7 +24,7 @@ static inline int write_all(int fd, const void *buf, size_t count) {
     while (bytes_written < count) {
         ssize_t res = write(fd, ptr + bytes_written, count - bytes_written);
         if (res <= 0) { if (errno == EINTR) continue; return -1; }
-        bytes_written += res;
+        bytes_written += (size_t)res;
     }
     return 0;
 }
@@ -35,7 +35,7 @@ static inline int read_all(int fd, void *buf, size_t count) {
     while (bytes_read < count) {
         ssize_t res = read(fd, ptr + bytes_read, count - bytes_read);
         if (res <= 0) { if (errno == EINTR) continue; return -1; }
-        bytes_read += res;
+        bytes_read += (size_t)res;
     }
     return 0;
 }
